+++ lib/pin_put/pin_put.dart
@@ -2,18 +2,6 @@ import 'package:flutter/material.dart';
 import './pin_put_view.dart';
 
 class PinPut extends StatefulWidget {
-  final bool isTextObscure;
-  final bool clearButtonEnabled;
-  final int fieldsCount;
-  final int clearButtonColor;
-  final double fontSize;
-  final double borderRadius;
-  final TextInputType keybaordType;
-  final TextInputAction keyboardAction;
-  final IconData clearButtonIcon;
-  final Function onSubmit;
-  final bool autoFocus;
-
   PinPut(
       {this.onSubmit,
       this.fieldsCount = 4,
@@ -23,10 +11,24 @@ class PinPut extends StatefulWidget {
       this.keybaordType = TextInputType.number,
       this.keyboardAction = TextInputAction.next,
       this.clearButtonIcon = Icons.backspace,
+      this.pasteButtonIcon = Icons.content_paste,
       this.clearButtonEnabled = true,
       this.clearButtonColor = 0xFF66BB6A,
       this.autoFocus = true})
       : assert(fieldsCount > 0 && borderRadius > 0);
+
+  final bool isTextObscure;
+  final bool clearButtonEnabled;
+  final int fieldsCount;
+  final int clearButtonColor;
+  final double fontSize;
+  final double borderRadius;
+  final TextInputType keybaordType;
+  final TextInputAction keyboardAction;
+  final IconData clearButtonIcon;
+  final IconData pasteButtonIcon;
+  final Function onSubmit;
+  final bool autoFocus;
 
   @override
   PinPutView createState() => PinPutView();