+++ lib/pin_put/pin_put_view.dart
@@ -10,25 +10,31 @@ class PinPutView extends PinPutViewModel {
   }
 
   Widget generateTextFields(BuildContext context) {
-    List<Widget> textFields = List.generate(widget.fieldsCount, (int i) {
+    List<Widget> widgets = List.generate(widget.fieldsCount, (int i) {
       return buildTextField(i, context);
     });
-    if (widget.clearButtonEnabled)
-      textFields.add(Padding(
-        padding: const EdgeInsets.only(left: 10.0),
-        child: GestureDetector(
-            onTap: () => clearWidget(context),
-            child: Icon(
-              Icons.backspace,
-              color: Color(widget.clearButtonColor),
-              size: 30.0,
-            )),
-      ));
+    if (widget.clearButtonEnabled) widgets.add(_actionButton());
     return Row(
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
         mainAxisSize: MainAxisSize.max,
         crossAxisAlignment: CrossAxisAlignment.center,
         verticalDirection: VerticalDirection.down,
-        children: textFields);
+        children: widgets);
+  }
+
+  Widget _actionButton() {
+    return StreamBuilder(
+      stream: bloc.onActionChanged,
+      initialData: false,
+      builder: (BuildContext context, AsyncSnapshot snap) {
+        return IconButton(
+            onPressed: () => onAction(context, snap.data),
+            icon: Icon(
+              snap.data ? widget.pasteButtonIcon : widget.clearButtonIcon,
+              color: Color(widget.clearButtonColor),
+              size: 30,
+            ));
+      },
+    );
   }
 }